refactor: split identity store to have prekey store be separate
